Hooks: Id column is db-generated; no longer maps to GitHub's id
The old plan was to have the id column map to GitHub's id for each hook.

As it turns out, we're going to want to insert a row for the hook into
our DB before we issue the create-hook call to GitHub.

We want to do this because GitHub's first webhook ping comes immediately
after creation, and when we handle the ping, we need to be able to look
up the hook row.  So, by writing the hook row first, we avoid the race.