enable de-duplication for AddOrUpdate(Repo|Org)Webhooks queues
Webhook installation has a race condition where simultaneously running
jobs could stomp on each other.  Also, it would just be wasteful to try
to install webhooks for the same repo or org concurrently.

The scenario looks like ---

1/ Are there any Hooks registered for repo X?  None?  Ok, installing.
2/ Fetch list of installed hooks from GitHub and remove any existing
   hooks that already point back to our host.
3/ Create Hook row for repo X, save to DB.
4/ Make request to GitHub to add a hook for repo X.
5/ Update Hook entry to include new webhook Id from GitHub.

There's a couple different ways the above could do the wrong thing if
we're running multiple jobs at the same time for the same repo or org.
So, we'll avoid that.

Closes #97 (or at least that's the hope)