webhooks: on repo deletion, handle case where repo is not org-owned
Previously this only considered the org-owned repo case.  But, in the
case where a repo is user-owned, we should trigger a sync for the repo
owner.

(We should also trigger a sync for contributors, but this doesn't do
that yet.  At least it's more correct than what we have now.)