webhooks: aggregate change summaries and fire notification once.
@kogir's feedback.

There was a case where, if we received an issue_comment event for an
issue that did not yet exist in our DB, we'd fire two change summaries.
The first would come from HandleIssues (as we created the issue), and
the second would come from HandleIssueComment (as we created the
comment).

We should aggregate and only fire once.