AccountOrganization gets its own Entity class.
In later commits, we're going to be adding a field (Admin) to the
AccountOrganizations join table.

EF does not really help you when you want to add extra fields to a join
table.  The only solution I can find online is to promote your join
table to its own Entity and manage the relationship yourself.