SyncOrganizationMembers knows if member is admin or not
GitHub's `/orgs/<name>/members` list does not include any role info for
the listed members.  The only way to determine role for each member is
to make two fetches to the endpoint: one with `?role=admin` and the
other with `?role=member`.

In a later commit, we'll start saving this role info to the DB.