send the ShipNeedsWebhookHelp flag when sync'ing orgs + repos
This flag should be true if 1) we have no registered hooks for the org
or repo, and 2) the user doing the sync is not an admin.

When true, the client will display a little warning prompting the user
to ask an Admin to login to ship (so we have rights to setup hooks).

We never set this flag for admin users because even if a hook doesn't
currently exist, it will get added very soon as part of the sync.