when webhook is added, trigger sync + change event for repo or org
Now that the RepositoryEntry and OrganizaitonEntry objects include the
ShipNeedsWebhookHelp flag, we need a way to toggle that flag to false
when a hook is added and we no longer need help.

So, we'll bump the row version every time we add a webhook to a repo or
org, and we'll trigger a change notification so any active clients will
re-sync.