fix #108 where we'd fail to unlabel if all labels were removed
Label deletion is done via a MERGE into the IssueLabels table.  The
deletion case would only remove issue-to-label associations if the
IssueId was present in the list of @Labels provided to BulkUpdateIssues.

In the case where all labels are being deleted, @Labels would be an
empty table and this condition was never met.

Instead, do the deletion if the IssueId is found in @Issues.