WebhookReaper cleans out "stale" Hooks that won't respond to pings
This is to catch the case where someone deletes or disables our webhook.

The WebhookReaper looks for Hook entries where the `LastSeen` field
looks pretty stale ("stale" right now means we have not seen anything
for this hook in >24 hours).

For hooks that look "stale", WebhookReaper fires off a ping and
increments the `PingCount` field on the hook.  After 3 pings and no
response, WebhookReaper deletes the Hook entry.

The Hook entry should get re-added the next time an admin of this repo
or org uses Ship (so a sync gets triggered).