when adding/updating webhooks, remove hooks where GitHubId is null
Part of the fix for #120.

When adding webhooks, there is always a chance that execution stops
AFTER we've created the place holder Hook entry, but BEFORE we receive
the success response from GitHub along with the hook's GitHub Id value.

In this case, we should consider any Hook with GitHubId == null as
invalid.  It should get removed, and we should try to add the webhook
again.