`/billing/accounts` provides info for app subscriptions window
A GET of `/biling/accounts` will provide a list of all accounts to be
shown in the client subscriptions window.  i.e., your own account and
any orgs you belong to.

For each account or org, an "action" field indicates what the allowable
action is for this account.  When action = purchase, show a Purchase
button.  The account does not have an existing subscription.  When
action = manage, show a Manage button.  The account has an existing
subscription.

Later we'll probably want to restrict the "manage" action to only the
person who bought the subscription or other org admins, so we'll need a
third action like "none_already_subscribed" or something.

This endpoint will return an error response if we've failed to sync the
subscription state from ChargeBee for all of these accounts.  If this
happens, we wouldn't know whether to show a manage or puchase option, so
let's just show nothing at all.