use Singleton to serialize queues; don't bother w/ dedup
I should have used this Singleton thing from the beginning.

The dedup solution with a 10-minute window (as it is right now), is a
pain.  e.g., when I'm testing the billing setup path, I open and close
the app repeatedly and inspect in ChargeBee to see if it worked.  With a
10-minute dedup window, I can only test once every 10 minutes for a
given user id.  Inside that 10 minute window, messages get dropped on
send.