you can buy a subscription for organizations
If this organization is buying a subscription for the first time, we use
ChargeBee's checkout-new-subscription page.

If the organization had a subscription for Ship in the past, we send the
user to ChargeBee's checkout-existing-subscription page.  That will give
them the opportunity to update billing info.

Unfortunately ChargeBee's checkout-existing-subscription page will not
re-activate their old subscription.  It stays in the cancelled state.

To workaround that, we have the checkout page redirect the user to the
/billing/reactivate endpoint on completion, and then we re-activate
their subscription via an API call.