close invoices on pending_invoice_created
We've enabled "Notify and wait to close invoices" mode on ChargeBee,
so we'll receive this webhook every time ChargeBee wants to invoice
someone.  This gives us an opportunity add on additional charges if we
want.  The charge won't actually happen until we close the invoice.

For now we just close all invoices w/out doing anything.  Later, we'll
add on extra charges for active users.