for organizations, we charge extra for >5 active users
Every month at subscription renewal time, ChargeBee will send us a
webhook event for "pending_invoice_created".  They create an invoice,
leave it in a pending state, and give us an oportunity to add extra
charges to it.

This pending invoice already includes the base org plan charge of $25,
and it's for the _coming_ month.  To this, we need to add the active
user charges for the _prior_ month.

On our side, we count up all the active users we've seen in the prior
month (i.e. [invoice date minus 1 month] through [invoice date minus 1
day], inclusive).  We then add the charge as an "addon" via ChargeBee
and close the invoice.