workaround chargebee bug in subscription_reactivated event
Backstory - When we tell ChargeBee to reactivate, we'll immediately get
a webhook event telling us the sub has been reactivated.  When we
receive that event with status = active, we change the subscription
state to Subscribed in the Subscription entity.

The "resource_version" value we get in the "subscription_reactivated"
event is bogus.  ChargeBee's working on a fix.  The resource version is
supposed to give us a timestamp of millis since 1970, and we use this in
the webhook handler to ignore events with out-dated info.

We can use another timestamp field in the webhook event to workaround,
but unfortunately there are no others with millis resolution - only
seconds.  To make this work, we'll delay reactivation by 1 second so
we're guaranteed that the "subscription_reactivated" event has a greater
timestamp than the "subscription_changed" event which comes in around
the same time.