UpdateComplimentarySubscription message handler
When a user is a member of a paying org, their Ship subscription is
complimentary.  We apply a coupon to discount it 100%.

We'll use this message handler to re-eval whether a user's subscription
should be complimentary, and we'll call it whenever that state might
change.

Nothing calls it yet - that's in the next couple commits.