Update and unify Service Bus code.
Turns out MessagingFactory.CreateQueueClient was a blocking call. This
was fine (but not great) until it was used in Orleans, where code is run
in a strict, single threaded scheduler. When the Service Bus code tried
to wait on a result, the async tasks it just launched never got to run
(single threaded scheduler). Thus is hung.

MessageSender lets us access queues and topics with the same code, and
properly suports async. Win!