outlook: fix issue where email address in preheader was visible.
The preheader text in our email is supposed to be invisible and only
show in message previews.

Our issue was that the preheader contained a mailto link referencing our
email address, and Outlook was allowing this link to display despite the
preheader style settings being "display: none".

To fix, we'll just exclude all the header+footer blurbs from the
preheader text so the email address cannot be included.