support for ignoring chargebee events; makes testing easier
Our ChargeBee test site (realartists-test) is setup to send webhook
events to shiphub-dev and my local machine, so webhook events can get
handled in two places.  That can be a problem when we act on the event
and make some change in ChargeBee state.  In that case, the event sent
to shiphub-dev might succeed while the event sent to my machine might
fail.

To workaround this, we can now configure shiphub-dev to ingore chargebee
events for a certain user.  And, we can configure our local machine to
only listen to events for a certain user.  No more conflicts.