fix bug where trial users belonging to paid org would go active
Fixes #194

Trials are supposed to auto-cancel at the end of 30 days unless you
provide payment info.

There was a bug where, if you belonged to a paying org, your trial would
automatically convert to "active".  As a member of a paying org, you got
a coupon for 100% off your subscription.  So, at the end of 30 days,
since your amount owed was $0 because of the coupon, ChargeBee would
automatically make you "active" instead of cancel you.

To fix, just don't add the coupon to users that are in trial.  We'll add
it later at purchase time if we need to.