handle webhooks for "label" events
We handle "created", "deleted", and "edited" events.

Tested by changing a label name in realartists/test via the web site.  On
the client, you can see the label drop down in the filter bar update
with the new label name.

The client doesn't know how to update label names/colors attached to
issues yet, but I think James is working on that.