Use a push webhook to watch for changes to ISSUE_TEMPLATE.
Avoid polling repo contents in RepositoryActor unless we have to, in
order to work around a GitHub bug where the cache details for these
endpoints are not stable for private repos. In the event that we must
poll for ISSUE_TEMPLATE (which happens when we cannot install a
webhook), poll only once every 5 syncs (5 minutes), in order to reduce
the impact on the rate limit.