Fix NPE - for some events, ChargeBee's customer record is not provided.
On "invoice_generated" and "pending_invoice_generated" events, ChargeBee
does not send us the "customer" in the hook payload, which means we
don't have the GitHub username handy.  To workaround, find the customer
id in other sections and lookup the GitHub login.

Note: None of this code runs in production as the
ChargeBeeWebhookIncludeOnlyList and ChargeBeeWebhookExcludeList app
settings will always be unset.

Closes #227