Fix OrleansAzureClient static initializer.
Unless I reference OrleansAzureUtils from code in a way that can't be
optimized away, it's not copied as a dependency. The old method no
longer works, because the AzureClient.IsInitialized accessor actually
reads CloudServices configuration now. This doesn't work when not in
CloudServices. The new hack (AppDirectoryLocations) never executes
unless called, and is never called.