don't run the ShouldIgnoreWebhook check for irrelevant events
The ShouldIgnoreWebhook check was barfing on some ChargeBee events that
did not contain customer info.  This is stuff like "plan_updated" which
fires whenever you change a plan in the ChargeBee admin site.
ShouldIgnoreWebhook barfs in this case because it can't identify which
customer this event references (it doesn't reference any).

We only care about a select number of event types.  To fix this error,
we'll delay the ShouldIgnoreWebhook check until after we're sure this is
an event we care about.