fix bug where emails would be sent even if ShouldIgnoreWebhook was true
Going back to calling ShouldIgnoreWebhook at the top of this function
and bailing out if it's false.

It used to be that way, but I changed it because ShouldIgnoreWebhook
would throw an exception when called for some ChargeBee events we don't
handle.  I've fixed that now so it's harmless to call it all the time.