Hack around Orleans serialization issue
Attempt to solve realartists/shiphub-server#247 by working around
dotnet/orleans#2528.

This patch intercepts all Orleans calls. If an exception is thrown, the
exception is checked for serializability. If it can be serialized, it is
re-thrown. If not, it's .ToString() representation is wrapped in a
standard exception.

Gross, but hopefully effective and not too slow.