Fix realartists/shiphub-server#290 Server returned 500 after creating a new issue
It seems like if the webhook or the repo polling gets in and creates a
new issue concurrent/before the GitHubProxyController can, that
BulkUpdateIssues can fail with a duplicate key error (as an aside,
it's important to remember that SQL MERGE doesn't guarantee that this
can't happen).

Regardless of all the reasons why updating our DB can fail, it's
important if GitHub has given a success response to pass that along to
the client.  If we don't, the client will think the issue hasn't been
filed and it may file a duplicate if the user requests it (which they
probably will).