/billing/accounts: return empty list instead of error
This is the endpoint that populates the Manage Subscriptions window.

There's an issue where -- if you were to do a fresh login, and
immediately open the Manage Subscriptions window, and if the server had
not yet managed to sync any subscription state from ChargeBee, we'd show
an error to the user.

You can repro by deleting all content in the Subscriptions table and
then opening Manage Subscriptions on the client.

Instead of showing an error, let's just return an empty list.  As soon
as the server syncs with ChargeBee, the window will automatically
refresh.