new subscription state is now immediately reflected after signup.
Fixes #268

Before this, our Subscriptions table was only updated via ChargeBee
webhook events.  When a user signed up, there was a period where their
subscription was active, but we didn't know it yet because we were
waiting for the ChargeBee webhook event to arrive.  Because of this, the
user might still see "Free Trial: XX days remaining" banner in the Ship
toolbar for a moment after signing up.

Now we'll immediately update the subscription state at the end of
signup.