On purchase, pass plan value and id to Thank You page
The Thank You page will use these values to record a "Purchase" pixel
event to FB and a "Goal" event to Google Analytics.

This allows us to bid on actual purchase events in our advertising
rather than clicks or downloads.  Though, I'm not actually sure if that
will work in the short-term with not many purchases - we'll see.