only use the "trial_days_left_{daysLeftOnTrial}" coupons
When I switched us to a 14-day trial, I created a whole new set of
coupons to be applied at purchase time for users that signup before the
trial ends.  Remeber, the purpose of these coupons is to credit users
that signup with trial time still remaining on the clock.

In the original 30 day trial scheme, I had created 30 coupons -
"trial_days_left_1" with a discount value of (1/30 = 3.3%) through
"trial_days_left_30" with a discount value of (30/30 = 100%).

When I switched us to 14 day trials, I created 13 new coupons -
"trial_days_left_1_of_14" with a discount value of (1/14 = 7.1%) through
"trial_days_left_14_of_14" with a discount value of (14/14 = 100%).

That was wrong and I never needed to do that.

If someone had signed up on the 1st day of their trial, I was giving
them the "trial_days_left_14_of_14" coupon with a 100% discount.  That's
not right - you should only get a 2 week credit for signing up on the
1st day of your trail, not a full-month credit.  Instead, I should have
just given them the "trial_days_left_14" coupon which gives a (14/30 =
43%) discount.

I realized the error soon after and updated the coupon discount
percentages on the ChargeBee side, but I left this code in here.  Not
sure if any real customers got a bigger discount than they were supposed
to, but it wasn't live that long (maybe a couple weeks?).  Time to clean
up the mess.